Fixed mmap usage for x86 Linux systems.
- huge pages are currently only used on Tilera systems (REMEMBER)
- Linux systems do not seem to map upwards like OS X (not only on Tilera)
- changed a few size related types from off_t to size_t, seems to be more consistent
- use _FILE_OFFSET_BITS = 64 to allow ftruncate to work with requests larger than 2GB-1 (REMEMBER, might be problematic on OS X, Tilera)

+++ vm/src/heap/memory_system.h
@@ -122,7 +122,7 @@ private:
   bool ask_Linux_for_huge_pages(int);
   int how_many_huge_pages();
   void request_huge_pages(int);
-  void  map_read_write_and_read_mostly_memory(int pid, int, int);
+  void  map_read_write_and_read_mostly_memory(int pid, size_t, size_t);
   void set_second_chance_cores_for_allocation(int);
 
 public:
